pyproject.toml
[project]
name = "mcpager"
version = "0.1.0"
description = "Pythonic MCP client"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.12.4",
    "langchain>=0.3.27",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "python-dotenv>=1.1.1",
]
